+++ tests/unit/core/keypairs/test_main.py
@@ -5,33 +5,31 @@ from xrpl.constants import CryptoAlgorithm
 from xrpl.core import keypairs
 from xrpl.core.keypairs.exceptions import XRPLKeypairsException
 
-_DUMMY_BYTES = b"\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x0c\r\x0e\x0f\x10"
+_DUMMY_HEX = "0102030405060708090a0b0c0d0e0f10"
 
 
 class TestMain(TestCase):
     # unfortunately, this patching is very brittle; it depends on the syntax
     # used to import secrets within the calling module.
     @patch(
-        "xrpl.core.keypairs.main.token_bytes", autospec=True, return_value=_DUMMY_BYTES
+        "xrpl.core.keypairs.main.token_bytes",
+        autospec=True,
+        return_value=bytes.fromhex(_DUMMY_HEX),
     )
     def test_generate_seed_no_params(self, _randbytes):
         output = keypairs.generate_seed()
         self.assertEqual(output, "sEdSKaCy2JT7JaM7v95H9SxkhP9wS2r")
 
     def test_generate_seed_entropy_provided(self):
-        output = keypairs.generate_seed(_DUMMY_BYTES.decode("UTF-8"))
+        output = keypairs.generate_seed(_DUMMY_HEX)
         self.assertEqual(output, "sEdSKaCy2JT7JaM7v95H9SxkhP9wS2r")
 
     def test_generate_seed_ed25519(self):
-        output = keypairs.generate_seed(
-            _DUMMY_BYTES.decode("UTF-8"), CryptoAlgorithm.ED25519
-        )
+        output = keypairs.generate_seed(_DUMMY_HEX, CryptoAlgorithm.ED25519)
         self.assertEqual(output, "sEdSKaCy2JT7JaM7v95H9SxkhP9wS2r")
 
     def test_generate_seed_secp256k1(self):
-        output = keypairs.generate_seed(
-            _DUMMY_BYTES.decode("UTF-8"), CryptoAlgorithm.SECP256K1
-        )
+        output = keypairs.generate_seed(_DUMMY_HEX, CryptoAlgorithm.SECP256K1)
         self.assertEqual(output, "sp5fghtJtpUorTwvof1NpDXAzNwf5")
 
     def test_derive_keypair_ed25519(self):